加入SQL語法

+++ eSaleModel/Order.cs
@@ -15,7 +15,7 @@ namespace eSaleModel
         public DateTime? RequiredDate { get; set; }
         public DateTime? ShippedDate { get; set; }
         public int ShipperID { get; set; }
-        public double Freight { get; set; }
+        public decimal Freight { get; set; }
         public string ShipName { get; set; }
         public string ShipAddress { get; set; }
         public string ShipCity { get; set; }
